Moved Server functionality to Traits (ProxmoxServer.php) and modified ProxmoxAuthentication to also be a trait instead of a class composite.
